+++ apps/frontend/src/components/ui/create-project-version/components/DependencySelect.vue
